requirements.txt
aiogram==3.15.0
groq==0.31.0
python-dotenv==1.1.0
PyMuPDF==1.25.3
openpyxl==3.1.5
